converting reads with homopolymer in last flow
If a read has a homopolymer at the last flow (nFlow == nFlows2) theres an error in the sff file as the loop get skipped and flow_index won't get updated to 0 for the homopolymers (leaving it to whatever value the array had on this position). 
This leads to flow indexes greater than the max number of flows.
Files with this error cann't be processed by newbler.

Removing the unnecessary query in the loop corrects this errors.